Make tooltips screen-wide/Fix tooltip z-index bug
